Split input and output streams into separate modules
