+++ crates/nu-cli/src/stream/mod.rs
@@ -1,0 +1,5 @@
+mod input;
+mod output;
+
+pub use input::*;
+pub use output::*;
